Flatten receive/uni.rs into receive/mod.rs
